Add specs testing our integration with gds-sso gem
To be able to use different auth providers in future we're going to
start replacing parts of the gds-sso gem on the fly, bringing some of
its functionality into our codebase while still relying on its code for
other parts. While we're doing this it would be good to be confident
that it's still working as intended; this commit adds some integration
specs that check the happy path.