Add event logger
We want our logs to have details formatted as JSON so it can be more
easily passed by Splunk. We're using lograge for the default Rails logs,
but for custom messages it's a bit of a pain to format the message by
hand. This commit adds an EventLogger to make it easier. The interface
is the same as what we have in forms-runner [[1]].

[1]: https://github.com/alphagov/forms-runner/blob/4fbf022e6c4618120bf32ce6228ce3ba0d54ba64/app/lib/event_logger.rb