Don't get user name from Auth0
For now we're using just the email address passwordless method for
authenticating users with Auth0. This means we're not guaranteed to get
a valid name from the service, instead sometimes we get the email
address in the name field instead, which is not what we want. So for now
we just don't ask for the name, as we don't need it for much, and maybe
in future we'll have to figure something out.