Fix spacing under status tag form task list page
When there are no questions there is unnecessary extra space under the
draft tag. This happens because we're including a paragraph with lots of
margin for the preview link, but there's no preview link (because there
are no pages to preview).

This commit changes the template to only render the container for the
preview link if preview link will be rendered.