Fix flaky spec with delete route confirmation
In our factory we were using faker to generate a 2 digit id.
In some instances it meant that more than one page could have
the same id. This was the issue with this flakey spec.

The method being tested was searching for the first id in the collection
that matched and returning that. In production this would work because
because pages could never have the same id.
```
form.pages.filter { |p| p.id == goto_page_id }.first&.question_text
```

With the test data in some instances it meant the method being tested
was finding the wrong page in the order. This was an error with the test data not the method being tested

eg. of when the tests failed
```
form.pages.map(&:id)
[62, 20, 34, 73, 34]
```
The method being tested i.e the `result` was returning the 3 item
in the collection but then comparing it to the last item which was
a completely different page.

The solution was instead of using faker to generate random id, it
would be better to use FactoryBot sequence method, where it automatically
increments with every instance built/created as part of the the spec.