rename SETTINGS__SERVICE_UNAVAILABLE to SETTINGS__MAINTENANCE_MODE_ENABLED
This is to help us understand that it relates to planned downtime instead
when a service is completely down and is not planned. If the service was
down, Rails server is not the best place to do this as Rails itself maybe
down. It  would be better to handle this outside Rails and let AWS handle
it.