Use user provider instead of default provider when user is signed in
When a user is signed in a key for their user record is stored in their
session cookie; this cookie is valid across app restarts and upgrades,
and also remains valid if the identity provider used by the app changes.

In testing we found that if a user was signed in and then the app was
subsequently re-configured to use a different identity provider, the
user profile/sign out links in the header would not be valid for them.

This commit fixes things by using the configured authentication provider
only at the point of signing in; currently signed in users will instead
see the links for the provider they are signed in with.

This also fixes an issue we noticed in production with signing out of
GOV.UK Signon; because our app's sign out route no longer needs a
provider parameter, it no longer collides with the `/auth/gds/sign_out`
route from the gds-sso gem, and no longer causes an infinite redirect
loop in the user's browser.