Fix test that was passing with false positive
The test "A trial user's forms move to their organisations on role
upgrade" was passing in error; previously the super admin user was
logged in for the whole test, because with the `mock_gds_sso` Warden
strategy `login_as` will not clear session data. As the super admin user
was also a member of test-org, the tests passed.

This commit stops the false positive by putting the super admin user in
a different organisation to the trial user, and restores the expected
behaviour by clearing session cookies in the middle of the test. This
revealed a few other issues with the test, which we also fix in this
commit.