Move exception rescue block to ApplicationController
We should handle all ActiveResource::ResourceNotFound
all the same across the app. Previously, this would have
only have handle this error raises from the home page
or the form show page.

The same issue with FormPolicy::UserMissingOrganisationError
error if the policy is called in other controllers this ends up
bubbling up and becoming a sentry alert.