Allow signing out of mock_gds_sso
It used to be that there wasn't any point to trying to sign out when
testing locally, as the `mock_gds_sso` provider would immediately sign
you back in again. However, when testing different authentication
providers locally and switching around it is nice to be able to see how
the transition might be for users.

This commit adds a simple redirect to the GOV.UK Signon integration
platform, mimicking the behaviour of the gds-sso gem itself (when not in
production configuration).