Refactor tests to use explicit organisation association override
Previously we were allowing tests to set `organisation_slug` on a user
and have that create an organisation association alongside. However, we
want to start using organisation IDs instead of slugs to refer to them
when using the forms API, so we should prefer to either set an explicit
organisation ID or assign an organisation with the factory, so that the
ID is not obscured.