Add organisations fetcher
Add a rake task for scraping the GOV.UK organisations API [[1]] and
saving the data we want to our database. This is based on what the
GOV.UK Signon team do themselves [[2]].

Note, I considered adding the [gds-api-adapters gem][3] for this, but in
the end I found I could implement the functionality we needed easily
enough using the Ruby standard library.

[1]: https://www.api.gov.uk/gds/gov-uk-organisations/#gov-uk-organisations
[2]: https://github.com/alphagov/signon/blob/95bd01cb7902e96b6303a274e9301d83e21a32b7/lib/organisations_fetcher.rb
[3]: https://github.com/alphagov/gds-api-adapters