Fix sign up action when not using auth0
Just calling `redirect_to_omniauth` is not the right way to send the
user to the sign in page; that action expects to be called after Warden
has done its thing, and will also redirect them to the page they were
trying to access, so in this case there would be an infinite redirect
loop. Instead do the redirect ourselves, sending them to the default
OmniAuth provider.