Refactor build files
The make file doesn't add anything to what Rails already has, and not conventional Rails.

The multitask run of rubocop with the -A (fix) flag which would change the code, and may or may not run before rspec. This makes the tests non-deterministic.

Running setup on every build slows down running tests. This should only be run once.

Running bundle-audit isn't a lint task, and it's already run as part of the GitHub Actions workflow.