Update README.md
Our README.md file in our app repos are (in my opinion) a bit messy and
could use some re-organising to make it the crucial information more
prominent and the rest easier to find.

This commit makes an attempt on the readme for this repo. The changes
include but are not limited to:

- Reorganise sections
- Combine duplicate sections on testing
- Add more details on running tests and linters
- Update section on updating Docker files
- Update Sentry section
- Update Notify section
- Update how to contribute section
- Update setting up the database section
- Add note about normalizing the locales
- Add note about Rakefile
- Remove references to PaaS, we're using AWS for production now
- Fix tests badge
- Fix link to .ruby-version file