Temporary Fix for Installing Chrome Driver
Chrome have changed where the latest version of Chrome Driver v115 is
published and the step in our test action to install ChromeDriver fails.

selenium-webdriver gem from version 4.10 can use v114 of
Chromedriver with version 115 of Chrome.

We will need to address this again since it is not expected to work with
v116 of Chrome when it is released.