Add paper_trail to User model
Currently we don't have a way to view the history of changes to a User,
so can't track why a user changed role without using logs for example.

This commit adds paper_trail version tracking to the User model.

Currently it only tracks a few attributes - role, has_access and
organisation_id.

The attributes are specified as we don't want to log every change which
comes in to our user model from the auth provider. The chosen attributes
are the ones super_admin users have control over, so are tracked.