Add rspec paper_trail and whodunnit
Add whodunnit information to paper_trail. This adds user to the versions
table to tack which user has made changes. The way versions track state
before changes means that it's a bit more subtle - there is more info on
the [papertrail
site](https://github.com/paper-trail-gem/paper_trail#4a-finding-out-who-was-responsible-for-a-change).

It's possible to set the current_user outside of requests for automated
tasks and console access - see the snippets at the papertrail wiki:
https://github.com/paper-trail-gem/paper_trail/wiki/Setting-whodunnit-in-the-rails-console

This commit also makes the papertrail helpers and matchers available in
tests.