Add AdvisoryLock helper
For some rake tasks we want to make sure that we don't run the task more
than once simultaneously. This commit adds an AdvisoryLock::DSL module
(similar to the VolatileLock::DSL module in alphagov/signon [[1]]) that
allows for acquiring locks. Underneath this uses the same mechanism that
prevents database migrations being run concurrently in ActiveRecord [[2]].

Note that it requires PostgreSQL. We are using this in our apps
currently, but if we used a different database acquiring the lock would
always fail and we would need to rewrite this code.

[1]: https://github.com/alphagov/signon/blob/95bd01cb7902e96b6303a274e9301d83e21a32b7/lib/volatile_lock.rb
[2]: https://github.com/rails/rails/blob/v7.0.6/activerecord/lib/active_record/migration.rb