Add the DFE autocomplete gem
We currently don't have an easy way for admin users to select from large
lists of choices.

This commit adds the GOV.UK accessible autocomplete component:

https://github.com/alphagov/accessible-autocomplete#api-documentation

This component is supported by the desgin system team but it isn't a
priority.

Given that we are using it only within the admin we are probably safe to
use it.

https://github.com/DFE-Digital/dfe-autocomplete

We ues the dfe-autocomplete package. The other option is to wrap the
accessible automplete component (which is JS and CSS) ourselves. There
doesn't seem to be much benefit to doing that, but it's an option if, in
the future the dependency on the DFE component becomes awkward.

Neither the Ruby Gem or node package have been published, both are
fetched from github. This means we now rely on Git to bundle and node
install. Luckliy we have just added Git to our docker images so there is
no new dependency.

- [x] Does it solve a specific problem we have?
        Yes, it packages up the accessible autocomplete component and
        allows us to easily integrate it into rails forms.
- [x] Is it updated regularly?
        Yes, it's had a couple of changes this year
- [x] Is it widely used?
        Doesn't seem to be used outside of DFE who have a handful of
        large projects which use it
- [x] Is it well documented?
        It's not bad - there isn't much to document but the readme is
        good
- [x] Does it have any major security issues?
        I don't think so
- [x] Would writing the code ourselves have a better outcome?
        We would be doing something very similar and probably wouldn't
        do it any better. We can always do it ourselves if it ever seems
          like we need to.

Reduce package-lock.json changes

Add DFE inflections