Add autocomplete component for organisation
When super admins set the user's organisation, there will be a long
(1000+) list of organisations to choose from.

Using a select list is not ideal for quickly changing the values.

This commit uses the autocomplete component to progressively enhance a
select component to a type-ahead style text input.

When the user does not have javascript available the component appears
as a select list. When JS is enabled an autocomplete component is used.

With the latest set of organisations, which contains about 1180
entries, the page size is about 90kb. This seems fine and choosing
organisations this way works, as long as you have an idea of the value
you need.

We need to allow the user to organisation to nil, as trial users who
have never had an organisation may need other values updating - e.g.
has_access might need to be set to false without setting an
organisation.

The autocomplete component doesn't naturally allow no value to be
entered into the progressively enhance version - the controller has to
take care to check the 'rawAttribute' value from the text input.

There are other options we could use on the autocomplete, if we decided
we need to:
- add synonyms for example DWP for Department of Work and Pensions
- show a list as well as allow typeahead

It's not clear how much value they would add, so they can be added as
needed.