Pin DfE autocomplete gem to commit ref
The DfE autocomplete gem has not been published and is only available
through github. Currently the dependency gets the latest main version
from github. If a failing build is committed or there are major changes
to it, our build will fail.

This commit pins the version to a specific commit.

There is a github release, but this is quite behind the latest code so
the commit was chosen instead.