Create page summary data service
We need to standardise how we build up data objects
that we use for passing to govuk_summary_list component.

Once this is in we should be able to refactor how
we return the list of answer settings summary list which
is currently a view component.

FYI - Ideally I only wanted to be able to pass in
the page to the service but because of the way our
ActiveResource / API urls are setup its not possible
to call something like `page.form` to make another api
call for the pages form. To get around this I decided that
we needed to pass in the form and the page that we would
like summary list for.