Add name to new auth0 users
Add name to the attributes passed to User#find_for_auth in the auth0
warden strategy.

This allows name to be set when new users are created using auth0.
Nickname rather than name is used as for us, using auth0 and
passwordless login, name is the full email address whilst nickname is
the string before the @ symbol.

The schema of the auth_hash is documented by omniauth:
https://github.com/omniauth/omniauth/wiki/Auth-Hash-Schema

The schema auth0 uses to provide the details in the above hash are
documented here:
https://auth0.com/docs/manage-users/user-accounts/user-profiles/normalized-user-profile-schema