Add logo for Auth0 widget to public dir
For trust and branding purposes we want the GOV.UK crown to appear on
our Auth0 login screens. Auth0 expects us to provide the logo via a URL.
This commit adds a copy of [govuk-mask-icon.svg] from GOV.UK Frontend to
our `public` folder, so it will be served from the server root.

Note that we could tell Auth0 to use the copy of this file already in
our `assets` folder, but that is fingerprinted by the Vite asset
pipeline with a content hash [[1]], so we could accidentally break our
logo when GOV.UK Frontend updates the crown logo by doing that.
t

[1]: https://vitejs.dev/guide/assets.html

[govuk-mask-icon]: https://github.com/alphagov/govuk-frontend/blob/v4.7.0/src/govuk/assets/images/govuk-mask-icon.svg