Sort users page by organisation, role, and name
Currently the users table that lists all users is not ordered. That’s ok
when there aren’t many users but it would probably make sense to order
them by organisation, name, user type or a combination of these - so
that it’s easier to find a specific user or get a view of a number of
users in a group.