Fix feature test using list of class instance instead of hash
While manually testing the work flow everything passed. However when running the capybara feature
specs, a list of Pages::SelectionOption were being saved
to session before trying to render the "Edit question"
page. This is a patch to handle this scenario or
to just use the hash vlaues that we passed to the
method to create new Pages::SelectionOption instances

We should be able to remove this once we replaced
form objects with normal activerecord resources