Add form for user to request editor upgrade
For early access, trial users need to be able to request an editor
account from the engagements team.

This requirement is met with a new page '/upgrade'. It contains some
text and a form with a confirmation box, which when filled in shows the
user a screen '/upgrade/requested' telling the user that they will
receive an email with further instructions.

This commit does not implement the email functionality but stubs out a
simple service which will provide this functionality.