fix issue with pages not being able to read/update from sessions
Recently, we did a refactor to convert all our session hash
strings to symbols for consistency reasons and for better performance.
https://github.com/alphagov/forms-admin/pull/637

A user reported issues with adding/editing guidance. On closer inspection,
this was because the session keys were symbols but the main `load_from session` method
was being passed in keys as strings rather than symbols.

` self.load("#{key}": session.dig(:page, key) || send(key.to_sym))`
This meant that session.dig was returning nil and therefore it was falling
back to whatever was last saved in forms-api.

The problem was not identify for a couple of reasons.

- The specs matched the calls being made in the controller so were all passing in strings.
- There was an issue with the spec that was written to check `load_from_session` returns the value from the session, had mocked out a session that was using strings as its hash key and so it was returning a false positive.
- There was also a missing spec to test what would happen if `load_from_session` was passed an array of strings