Add mailer and logging for upgrade requests
Add sending of email for user_upgrade_request and logging

When a user requests to upgrade from a trail user to an editor, they
should receive an email with more instructions.

The email should have a reply-to of our teams zendesk account so that
when the user replies, the answer is passed on to the engagement team.

An event is also logged.