Create new QuestionForm object
This will contain all the attribute that the user will
need to complete to create a question. Currently, it doesn't
fully integrate into the application as we arent storing these
values in the session nor do we have an activerecord to store
the data in. This will be done in a future commit.