Add metrics summary component
Adds a new component for displaying the metrics summary and its error states, and renders the component on the live form view if the feature flag is enabled.

Also adds tests and component preview for the new component.

Incorporates PR suggestions to use Ruby 'ago' syntax and govuk-frontend's govuk-spacing and $govuk-border-width helpers.

Co-authored-by: Alistair Laing <aliuk2012@users.noreply.github.com>