Replace Pages::SelectionOption with a simple Hash
`Pages::SelectionOption`inherits from `BaseForm` form object class which
makes it an ActiveModel and includes ActiveModel validations. Validations
are not needed to be included and we cannot stop this class from inherit
BaseForm without breaking a lot of things.

Instead we should replace the answer_settings selection_options to just be
a simple hash but for us to use it in the GOV.UK Design System form builder
we need to make that hash into an OpenStruct. Theres no need at this time to
abstruct and complicate how we build up these select list options.