Fix browserslist config to prevent unwanted Autoprefixer styles
This browserslist query includes IE11, which does not support CSS grid but does support a vendor prefixed version with a different API. This means that when we use CSS grid syntax, Autoprefixer will automatically generate CSS using IE11's vendor prefixed grid syntax. This can cause broken layouts when we rely on CSS grid's automatic placement, which isn't supported in the IE11 implementation.
In this commit we add 'and not dead' to the query to exclude unsupported browsers like IE11.