Update lint task to run sequentially
Runs the JS and style linting sequentially instead of in parallel. Running them in parallel meant that the linting tasks wouldn't break the build unless both lints failed, so we saw issues being missed. The equivalent was done in forms-runner[1] .

The time saving from running them in parallel is negligible, so there shouldn't be any noticeable drawbacks to changing this.

[1]: https://github.com/alphagov/forms-runner/pull/412