Revert "Refactor predicate methods to simplify"
This reverts commit 2a83c29dfa3db3bc932a328a70208d6066c5f0c7.

The refactor of the validation predicates in commit 2a83c29 caused
unexpected side-effects. Although in most cases checking the current
role is equivalent to checking whether the role has been changed, the
two conditions are not exactly identical. We found that after applying
this change we started getting errors from GOV.UK Signon because the
Signonotron API user had an editor role and no organisation set.

Although the original logic was more complex, it more precisely
describes the behaviour we want and has fewer bugs, so let's revert to
that. This commit also adds tests to show the behaviour we want.