Refactors answer_settings selection options
Attempting to refactor how we setup the selection
answer types selection_options. By default, it should
be treated as a hash but what complicates things at the
moment is that we store/retrieve it from various sources i.e
from session, ActionController::Params & forms-api (using ActiveResource).

I had issues with this refactor because sometimes the tests
were passing but the app was failing and visa-versa. It
ended up that the tests data was incorrectly setup and as
you can see a number of spec had to have answer_settings
defined closer to specs rather than in factorybot.

We can continue to tidy this up further but I think we need
to take baby steps and slowly improve how we define and use
answer_settings (more specifically "selection" answer type
which has deeply embedded hashes)