Find or initialize DraftQuestion
We need to create a new DraftQuestion instance/record
if the user is creating a new question.

If the user is editing an existing question, we try to
either find an existing DraftQuestion record which may
not have been fully saved to forms-api or we create a new
DraftQuestion and default its values to what Page has.