Add MouSignature model, migration, and specs
We want to record that form creators have signed an MOU on behalf of
their organisation.

In early access, agreeing to an MOU will be managed by the engagement
team. This will change when we move to self-signup. The current MOU work
does not prevent form creators from making forms live, doesn't add a
step to the task list and the journey to agree to the MOU will be
initiated by the engagement team.

The requirements for MOU signature are:
- trial users and editors can sign an MOU
- more than one user belonging to an organisation can sign an MOU

This commit adds a new model called `MouSignature` along with its
associated migration file and specs.

Allowing trial users to sign an MOU adds a difficulty - trial users are
not associated with an organisation. To get round this, MouSignature
does not require organisation. Trial users can still sign an MOU and
orgainsation will be set to nil, when the user is associated with an
organisation at a later point, the MouSignature will be modified to add
the organisation.

The `MouSignature` model belongs to a user and an organization
(optional). It also has a method `update_user_orgs` which updates
`MouSignature` records that have a `user` but no `organization`.

The migration file creates a table for `MouSignatures` with columns:
`user_id`, `organisation_id`, `agreed_at`, `created_at`, and
`updated_at`. It also adds indexes for the `user_id`, `organisation_id`
(unique), and `user_id` where `organisation_id` is null (unique). The
indexes prevent a user signing more than on MOU for a single
organisation and stops them signing more than one MOU when they don't
have an organisation set.

The spec file tests the validations and constraints of `MouSignature`
model.

It's possible to use Rails acronym inflections to use MOUSignature
instead of MouSignature instead. I decided it wasn't worth it though.