Add mou_signature_policy
Only super_admins will be able to list the MouSignatures.

This commit adds a new policy without a scope to use in future commits
to choose weather a user can view the signed MOUs.

Add a new file `mou_signature_policy.rb` to the `app/policies`
directory. This file contains a class `MouSignaturePolicy` with a method
`can_manage_mous?` that checks if the user is a super admin.

Also, add a corresponding test file `mou_signature_policy_spec.rb` to
the `spec/policies` directory. This file tests the `MouSignaturePolicy`
class with different user roles, allowing actions for super admins and
forbidding actions for other roles.