Vary heading text depending on form live date
Adds logic to vary the heading text depending on when the form went
live. If the form went live today, we don't show any date information or
the number of days in the heading. If the form went live yesterday, we
just show yesterday's date. If the form went live before that, we show
the date range and the number of days covered in the heading.