Remove form_is_new from metrics_data response
The metrics summary component has enough information to determine
whether the form is new, so we don't need to pass it this information.
This PR removes it from the metrics_data response, though it's still
used within the metrics_data response to stop the CloudWatch request
being sent for a new form.