Add check that form is live
Adds a check to the metrics_data method to make it return a nil object
in case the method is called on a form that isn't live yet. There
shouldn't be any way for this to happen in the app at present, but
guarding against it should make this method more robust.