Remove unnecessary live check
We added a check to see if the form is live, but this is problematic for
two reasons. Firstly, the has_live_version property it uses isn't
present in the live form JSON, so it will cause an error. Secondly, if
this method is called on a non-live form, that indicates that the
application has gone wrong, so we should cause an error rather than
handle it silently.