Add provider to User when using basic_auth
When using the basic auth authentication, we don't set provider on the
User object.

Our other auth methods - signon, auth0 etc do. This commit adds it and
sticks a tests in to check its there too.

This allows us to rely on provider being present in other parts of the
code.