Remove obsolete table called sessions
This table may have been setup during a spike into how we were going to built draft/live versions in forms-admin. We have never used it
but some local setups still have it in the db
and when db:migrate is run it gets added and
committed back to the repo.