Correct message about URL staying the same
When you make the draft of a live form live, you were told:

“The URL for the live form will remain the same.“

However, the URL may be different if the form name has changed. Links to the form will still work in this case, as they rely on the first part of the URL which never changes.

So we are changing this line to be more accurate and specific - to make sure people don’t worry unnecessarily about links to their form breaking.