Add default value for classes parameter
Adds a default value of `[]` to the NavigationItem struct's classes
parameter. This prevents a 'classes is nil, if no custom classes are
needed omit the param' warning from the govuk-components gem. This also
forces us to pass keyword arguments to the constructor, which is more
explicit and less prone to human error than using positional arguments.