Make accessibility tests more targeted
Our automated accessibility tests check all of the links on the /preview
page for accessibility issues. However, the selector for links is not
specific enough, so it was including the links in the header and footer,
meaning we were unintentionally testing these pages as part of the
component tests. This resulted in an error when we added the support
page to the header, since this link runs in another app - so it will
cause the tests to fail if the other app is not running.