Add failure page with correct retry link
Currently we're using the failure page from gds-sso gem [[1]] if there
is an issue when authenticating with OmniAuth. However, the redirect
link in that failure page is hardcoded to send users to GOV.UK Signon,
which isn't always what we want.

This commit copies the failure page into our own app so we can edit it,
and makes the link dependant on the `strategy` parameter from OmniAuth
(which should match the provider, at least in the scenarios we care
about).

[1]: https://github.com/alphagov/gds-sso/blob/3aa846de2d71c1f90c120546ade3595e0883c60c/app/views/authentications/failure.html.erb