Add UserUpdateService
When the user is updated through the users_controller we look for
specific changes so we can do extra tasks. Currently we check for a user
being upgraded from a trial user and modify the forms they own.

Adding MOUs means that we also need to check that a user has been
asigned an organisation so that the MOU organisation can be set too.

This commit extacts these actions into a service which can be called
from the controller to make it easier to test the behaviour of each.

The tests request tests for the users_controller can then be simplified
as they don't need to check for so many conditions.