Add routes, controller, locales for mou_signature
We need to allow users to sign MOUs.

This commit adds the views and controller actions to allow users to
agree to an MOU.

A single MOU can be signed by a user with any role, whether they have an
organisation set or not.

MOU's created without an organisation will have it set when the user is
given an organisation.