Convert basic auth user to be a trial user by default
By changing the users from editor to trial account. it meant that
the mock calls to form-api for the list of forms changed from returning
a list of forms by organisation to a list of forms. By doing this I
had to create a user upfront to have a creator id to add to the mock url.

As a result of using factorybot for creating a user, i had to update the
the email address but this shouldn't be a problem as we dont really use
it for anything