consistently set default answer settings
when "adding" a new question, we set the answer_settings to nil.
While setting a new question if a user goes back and changes the
answer type (or even just saving it without changing the answer type)
they will lose all their existing answer_settings. This is a bug
and we will come back to it in the next commit.

When "editing" an existing question, if a user changes an answer_type
we would then set the answer_settings to a default for that answer type.
Instead of setting the answer_settings in the form object this was
being done in the controller so it was harder to spot this mutation
of the value.

As of this commit we should be setting default answer_Settings in the form object