Save time in docker build by skipping vite:install_dependencies
The `assets:precompile` Rake task is defined by [vite-ruby] to run
`vite:install_dependencies` and then `vite:build_all` [[1]]. This is
fine, except that the task `vite:install_dependencies` runs
`npm ci` [[2]], which takes a while to do.

Since we've done this already earlier in the Dockerfile, we can skip
this step and go straight to `vite:build_all`. On my machine this saves
around 2 minutes from the build time.

[1]: https://github.com/ElMassimo/vite_ruby/blob/vite_ruby%403.3.4/vite_ruby/lib/tasks/vite.rake#L66-L67
[2]: https://github.com/ElMassimo/vite_ruby/blob/vite_ruby%403.3.4/vite_ruby/lib/tasks/vite.rake#L46