Add link to mou_signature from index
The app does not have a link to the mou_signature page. This is on
purpose: the engagement team will share the URL with teams who need it.

This commit puts the link in the super_admin mou menu item. This makes
it easier for the engagement team to acess the URL.