Show notification banner when role is upgraded to editor
This commit adds a banner which is shown to users when they first go
to the index after their role is upgraded to editor. This uses the
PaperTrail version history to establish if the role has changed to
editor, and will create a new version the first time the role changes
to editor, to prevent the banner being shown subsequently.