Pages::QuestionForm does not require form or page id
Theres no need for the form object to know about form id or page id. Its primary forcus is on display a question text, hint text and is optional field. Behind the scenes it still need answer type because in the view answer type is used as part of the locales key lookup to customise the hint text for the hint text field.