Display the information we have in draft question to the user
I discovered that if you added a detailed guidance existing question or edit guidance from existing question. The data would not show up on the "Edit Question page". This was due to a recent refactor and as we had no tests covering that part of the functionality it was not picked up.

This is a temporary fix, ideally once we add the "Check your question" page we will be able to completely rely on draft questions to display the current state to the user. Until then we will copy what we have in draft question over to the page model.

Caused by c9541b3e0bed7820abacd5bc8089245124f43917